IGNITE-25057 Move GridClientByteUtils logic to IgniteUtils
