+++ modules/clients/src/test/java/org/apache/ignite/internal/processors/rest/TestBinaryClient.java
@@ -44,7 +44,7 @@ import org.apache.ignite.internal.processors.rest.client.message.GridClientTaskR
 import org.apache.ignite.internal.processors.rest.client.message.GridClientTaskResultBean;
 import org.apache.ignite.internal.processors.rest.client.message.GridClientTopologyRequest;
 import org.apache.ignite.internal.processors.rest.protocols.tcp.GridMemcachedMessage;
-import org.apache.ignite.internal.util.GridClientByteUtils;
+import org.apache.ignite.internal.util.IgniteUtils;
 import org.apache.ignite.internal.util.typedef.F;
 import org.apache.ignite.internal.util.typedef.internal.U;
 import org.apache.ignite.logger.java.JavaLogger;
@@ -189,9 +189,9 @@ final class TestBinaryClient {
 
                                     GridClientResponse msg = marsh.unmarshal(msgBytes);
 
-                                    long reqId = GridClientByteUtils.bytesToLong(hdrBytes, 0);
-                                    UUID clientId = GridClientByteUtils.bytesToUuid(hdrBytes, 8);
-                                    UUID destId = GridClientByteUtils.bytesToUuid(hdrBytes, 24);
+                                    long reqId = IgniteUtils.bytesToLong(hdrBytes, 0);
+                                    UUID clientId = IgniteUtils.bytesToUuid(hdrBytes, 8);
+                                    UUID destId = IgniteUtils.bytesToUuid(hdrBytes, 24);
 
                                     msg.requestId(reqId);
                                     msg.clientId(clientId);
